Mirror the image demo-source layout for audio, video and text
Audio, video and text each inlined the demo-dataset catalog and the
per-source download+embed dispatcher in media_type.py, the way image
did before it was split. Move each type's demo machinery into a
sibling _demo_sources.py and leave media_type.py delegating, matching
vtscore/media/image/_demo_sources.py.

Also promote the per-category slice loop, written identically in
image, audio and video, to demo_slice_by_category() in
vtscore/media/base.py beside demo_slice(). Text's variant returns
parallel (texts, categories) lists rather than one flat list, so it
keeps its own loop.

Pure moves; no behaviour change. Audio and video's loaders take the
calling MediaType as their first argument because they need
load_media_data() for each clip's duration and thumbnail.
